restored schema-compilers file
