+++ lib/schema-controller.js
@@ -1,10 +1,10 @@
 'use strict'
 
 const { Schemas } = require('./schemas')
-
-const Ajv = require('ajv')
-const fastJsonStringify = require('fast-json-stringify')
-
+const {
+  ValidatorSelector,
+  SerializerCompiler: buildDefaultSerializer
+} = require('./schema-compilers')
 let buildPerformanceValidator
 
 module.exports = SchemaController
@@ -116,59 +116,4 @@ SchemaController.prototype.setupSerializer = function (serverOption) {
     return
   }
   this.serializerCompiler = buildDefaultSerializer(this.schemaBucket.getSchemas())
-}
-
-function buildDefaultSerializer (externalSchemas) {
-  return function ({ schema, method, url, httpStatus }) {
-    return fastJsonStringify(schema, { schema: externalSchemas })
-  }
-}
-
-function ValidatorSelector () {
-  const validatorPool = new Map()
-  const cache = new Map()
-  cache.put = cache.set
-
-  return function buildCompilerFromPool (externalSchemas, options) {
-    const externals = JSON.stringify(externalSchemas)
-    const ajvConfig = JSON.stringify(options.customOptions)
-
-    const uniqueAjvKey = `${externals}${ajvConfig}`
-    if (validatorPool.has(uniqueAjvKey)) {
-      return validatorPool.get(uniqueAjvKey)
-    }
-
-    const compiler = ValidatorCompiler(externalSchemas, options, cache)
-    validatorPool.set(uniqueAjvKey, compiler)
-
-    return compiler
-  }
-}
-
-function ValidatorCompiler (externalSchemas, options, cache) {
-  // This instance of Ajv is private
-  // it should not be customized or used
-  const ajv = new Ajv(Object.assign({
-    coerceTypes: true,
-    useDefaults: true,
-    removeAdditional: true,
-    // Explicitly set allErrors to `false`.
-    // When set to `true`, a DoS attack is possible.
-    allErrors: false,
-    nullable: true
-  }, options.customOptions, { cache }))
-
-  if (options.plugins && options.plugins.length > 0) {
-    for (const plugin of options.plugins) {
-      plugin[0](ajv, plugin[1])
-    }
-  }
-
-  for (const schemaId in externalSchemas) {
-    ajv.addSchema(externalSchemas[schemaId])
-  }
-
-  return function ({ schema, method, url, httpPart }) {
-    return ajv.compile(schema)
-  }
 }
